dashboard/requirements.txt
dash>=2.16.0
dash-bootstrap-components>=1.6.0
plotly>=5.22.0
pandas>=2.2.0
python-dotenv>=1.0.0
snowflake-connector-python>=3.10.0
